- Remove unwanted example files from TV target - Register Kotlin src dirs with Android plugin for AS IDE to work properly - Fix unit tests
