+++ Code/Client/Android/common/src/main/res/values/strings.xml
@@ -1,3 +1,8 @@
 <resources>
     <string name="app_name">bm_client_android_common</string>
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+
+    <string name="call_to_action_text">Enter your name</string>
+    <string name="response_placeholder">Server will respond here</string>
+    <string name="name_hint">Your name</string>
 </resources>
