+++ Code/Client/Android/common/src/test/kotlin/example/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
-package example.common
+package example
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
