+++ Code/Client/Android/tv/src/main/AndroidManifest.xml
@@ -29,9 +29,6 @@
                 <category android:name="android.intent.category.LEANBACK_LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity android:name="example.DetailsActivity"/>
-        <activity android:name="example.PlaybackActivity"/>
-        <activity android:name="example.BrowseErrorActivity"/>
     </application>
 
 </manifest>
