--- Code/Client/Android/tv/src/main/kotlin/example/BrowseErrorActivity.kt
@@ -1,84 +1,0 @@
-/*
- * Copyright (C) 2017 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-package example
-
-import android.app.Activity
-import android.app.Fragment
-import android.os.Bundle
-import android.os.Handler
-import android.view.Gravity
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.FrameLayout
-import android.widget.ProgressBar
-
-/**
- * BrowseErrorActivity shows how to use ErrorFragment.
- */
-class BrowseErrorActivity : Activity() {
-
-    private lateinit var mErrorFragment: ErrorFragment
-    private lateinit var mSpinnerFragment: SpinnerFragment
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        testError()
-    }
-
-    private fun testError() {
-        mErrorFragment = ErrorFragment()
-        fragmentManager
-            .beginTransaction()
-            .add(R.id.main_browse_fragment, mErrorFragment)
-            .commit()
-
-        mSpinnerFragment = SpinnerFragment()
-        fragmentManager
-            .beginTransaction()
-            .add(R.id.main_browse_fragment, mSpinnerFragment)
-            .commit()
-
-        val handler = Handler()
-        handler.postDelayed({
-            fragmentManager
-                .beginTransaction()
-                .remove(mSpinnerFragment)
-                .commit()
-            mErrorFragment.setErrorContent()
-        }, TIMER_DELAY)
-    }
-
-    class SpinnerFragment : Fragment() {
-        override fun onCreateView(
-            inflater: LayoutInflater, container: ViewGroup?,
-            savedInstanceState: Bundle?
-        ): View? {
-            val progressBar = ProgressBar(container?.context)
-            if (container is FrameLayout) {
-                val layoutParams = FrameLayout.LayoutParams(SPINNER_WIDTH, SPINNER_HEIGHT, Gravity.CENTER)
-                progressBar.layoutParams = layoutParams
-            }
-            return progressBar
-        }
-    }
-
-    companion object {
-        private val TIMER_DELAY = 3000L
-        private val SPINNER_WIDTH = 100
-        private val SPINNER_HEIGHT = 100
-    }
-}
